Added being able to rate tutors and being able to withdrawl to see_session

+++ app/src/main/java/com/example/project/reg_tutor.java
@@ -178,7 +178,7 @@ public class reg_tutor extends AppCompatActivity {
                                     map.put("Job", "Tutor");
                                     map.put("Status", 0);
                                     map.put("Rating",0);
-                                    map.put("NumberOfRaitings", new  ArrayList<Integer>());
+                                    map.put("NumberOfRaitings", new  ArrayList<Long>());
